% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antiread}
\alias{antiread}
\title{Antiread dataset}
\usage{
data(antiread)
}
\description{
Data for random coefficients regression, with smiles as dependent
variable, beers as independent variable, and bars as cluster
}
\examples{
data(antiread)
}
\keyword{datasets}
