% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clpm.h.R
\name{clpm}
\alias{clpm}
\title{Cross lagged panel model}
\usage{
clpm(
  data,
  factors = list(list(label = "X", vars = list()), list(label = "Y", vars = list())),
  constrain_crosslagged = FALSE,
  constrain_autoregressions = FALSE,
  constrain_latent_variances = FALSE,
  constrain_observed_errors = FALSE,
  constrain_covariances = FALSE,
  estimate_observed_intercepts = TRUE,
  estimate_observed_errors = FALSE,
  estimate_latent_intercepts = FALSE,
  show_lavaan_syntax = FALSE,
  missing_data_treatment = "listwise",
  show_schematic_plot = TRUE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{factors}{a list containing named lists that define the \code{label}
of the factor and the \code{vars} that belong to that factor}

\item{constrain_crosslagged}{.}

\item{constrain_autoregressions}{.}

\item{constrain_latent_variances}{.}

\item{constrain_observed_errors}{.}

\item{constrain_covariances}{.}

\item{estimate_observed_intercepts}{.}

\item{estimate_observed_errors}{.}

\item{estimate_latent_intercepts}{.}

\item{show_lavaan_syntax}{.}

\item{missing_data_treatment}{.}

\item{show_schematic_plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$lavaan_warnings} \tab \tab \tab \tab \tab a table \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$lavaan_syntax} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$schematic_plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$lavaan_warnings$asDF}

\code{as.data.frame(results$lavaan_warnings)}
}
\description{
Cross lagged panel model
}
