% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riclpm.h.R
\name{riclpm}
\alias{riclpm}
\title{Random Intercept CLPM}
\usage{
riclpm(
  data,
  factors = list(list(label = "X", vars = list()), list(label = "Y", vars = list())),
  constrain_crosslagged = TRUE,
  constrain_autoregressions = TRUE,
  estimate_intercepts_intercepts = TRUE,
  estimate_observed_intercepts = FALSE,
  constrain_covariances = FALSE,
  show_lavaan_syntax = FALSE,
  missing_data_treatment = "listwise"
)
}
\arguments{
\item{data}{the data as a data frame}

\item{factors}{a list containing named lists that define the \code{label}
of the factor and the \code{vars} that belong to that factor}

\item{constrain_crosslagged}{.}

\item{constrain_autoregressions}{.}

\item{estimate_intercepts_intercepts}{.}

\item{estimate_observed_intercepts}{.}

\item{constrain_covariances}{.}

\item{show_lavaan_syntax}{.}

\item{missing_data_treatment}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$lavaan_warnings} \tab \tab \tab \tab \tab a table \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$lavaan_syntax} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$lavaan_warnings$asDF}

\code{as.data.frame(results$lavaan_warnings)}
}
\description{
Random Intercept CLPM
}
